redis - feat: (breaking) upgrading @redis/client to v5
