+++ packages/redis/package.json
@@ -41,7 +41,7 @@
 	},
 	"homepage": "https://github.com/jaredwray/keyv",
 	"dependencies": {
-		"@redis/client": "^1.6.0",
+		"@redis/client": "^5.6.0",
 		"cluster-key-slot": "^1.1.2",
 		"hookified": "^1.10.0"
 	},
