+++ packages/redis/test/test.ts
@@ -5,127 +5,7 @@ import {createClient, type RedisClientType} from '@redis/client';
 import {delay} from '@keyv/test-suite';
 import KeyvRedis, {createKeyv} from '../src/index.js';
 
-describe('KeyvRedis', () => {
-	test('should be a class', () => {
-		expect(KeyvRedis).toBeInstanceOf(Function);
-	});
 
-	test('should have a client property', () => {
-		const keyvRedis = new KeyvRedis();
-		expect(keyvRedis.client).toBeDefined();
-	});
-
-	test('should be able to create Keyv instance', async () => {
-		const keyv = createKeyv('redis://localhost:6379', {namespace: 'test'});
-		expect(keyv).toBeDefined();
-		expect(keyv.namespace).toBe('test');
-		expect(keyv.store.namespace).toBe('test');
-		await keyv.set('mykey', 'myvalue');
-		await keyv.set('mykey2', {foo: 'bar'});
-		const value = await keyv.get<string>('mykey');
-		expect(value).toBe('myvalue');
-		// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-		const value2 = await keyv.get('mykey2');
-		expect(value2).toEqual({foo: 'bar'});
-	});
-
-	test('should be able to set the client property', () => {
-		const keyvRedis = new KeyvRedis();
-		const client = createClient() as RedisClientType;
-		keyvRedis.client = client;
-		expect(keyvRedis.client).toBe(client);
-	});
-
-	test('should be able to pass in a client to constructor', () => {
-		const client = createClient() as RedisClientType;
-		const keyvRedis = new KeyvRedis(client);
-		expect(keyvRedis.client).toBe(client);
-	});
-
-	test('should be able to pass in client options to constructor', () => {
-		const uri = 'redis://foo:6379';
-		const keyvRedis = new KeyvRedis({url: uri});
-		expect((keyvRedis.client as RedisClientType).options?.url).toBe(uri);
-	});
-
-	test('should be able to pass in the url and options to constructor', () => {
-		const uri = 'redis://localhost:6379';
-		const keyvRedis = new KeyvRedis(uri, {namespace: 'test'});
-		expect((keyvRedis.client as RedisClientType).options?.url).toBe(uri);
-		expect(keyvRedis.namespace).toBe('test');
-	});
-
-	test('should be able to pass in the url and options to constructor', () => {
-		const uri = 'redis://localhost:6379';
-		const options = {
-			namespace: 'test',
-			keyPrefixSeparator: '->',
-			clearBatchSize: 100,
-			useUnlink: true,
-			noNamespaceAffectsAll: true,
-		};
-		const keyvRedis = new KeyvRedis(uri, options);
-		expect(keyvRedis.namespace).toBe('test');
-		expect(keyvRedis.keyPrefixSeparator).toBe('->');
-		expect(keyvRedis.clearBatchSize).toBe(100);
-		expect(keyvRedis.useUnlink).toBe(true);
-		expect(keyvRedis.noNamespaceAffectsAll).toBe(true);
-	});
-
-	test('should be able to get and set properties', () => {
-		const keyvRedis = new KeyvRedis();
-		keyvRedis.namespace = 'test';
-		keyvRedis.keyPrefixSeparator = '->';
-		keyvRedis.clearBatchSize = 1001;
-		keyvRedis.useUnlink = false;
-		expect(keyvRedis.namespace).toBe('test');
-		expect(keyvRedis.keyPrefixSeparator).toBe('->');
-		expect(keyvRedis.clearBatchSize).toBe(1001);
-		expect(keyvRedis.useUnlink).toBe(false);
-	});
-
-	test('keyPrefixSeparator should be able to set to blank string', () => {
-		const keyvRedis = new KeyvRedis('redis://localhost:6379', {keyPrefixSeparator: ''});
-		expect(keyvRedis.keyPrefixSeparator).toBe('');
-		keyvRedis.keyPrefixSeparator = '->';
-		expect(keyvRedis.keyPrefixSeparator).toBe('->');
-		keyvRedis.keyPrefixSeparator = '';
-		expect(keyvRedis.keyPrefixSeparator).toBe('');
-	});
-
-	test('clearBatchSize should not set if 0 or less than', () => {
-		const keyvRedis = new KeyvRedis('redis://localhost:6379', {clearBatchSize: 0});
-		expect(keyvRedis.clearBatchSize).toBe(1000);
-		keyvRedis.clearBatchSize = 200;
-		expect(keyvRedis.clearBatchSize).toBe(200);
-		let error = '';
-		keyvRedis.on('error', message => {
-			error = message as string;
-		});
-		keyvRedis.clearBatchSize = -1;
-		expect(error).toBe('clearBatchSize must be greater than 0');
-		expect(keyvRedis.clearBatchSize).toBe(200);
-	});
-
-	test('should be able to get and set opts', async () => {
-		const keyvRedis = new KeyvRedis();
-		keyvRedis.opts = {
-			namespace: 'test', keyPrefixSeparator: ':1', clearBatchSize: 2000, noNamespaceAffectsAll: true,
-		};
-
-		expect(keyvRedis.opts).toEqual({
-			namespace: 'test',
-			keyPrefixSeparator: ':1',
-			clearBatchSize: 2000,
-			dialect: 'redis',
-			url: 'redis://localhost:6379',
-			noNamespaceAffectsAll: true,
-			throwErrors: false,
-			throwOnConnectError: true,
-			useUnlink: true,
-		});
-	});
-});
 
 describe('KeyvRedis Methods', () => {
 	beforeEach(async () => {
@@ -276,133 +156,7 @@ describe('KeyvRedis Methods', () => {
 	});
 });
 
-describe('KeyvRedis Namespace', () => {
-	beforeEach(async () => {
-		const keyvRedis = new KeyvRedis();
-		const client = await keyvRedis.getClient() as RedisClientType;
-		await client.flushDb();
-		await keyvRedis.disconnect();
-	});
-	test('should clear with no namespace', async () => {
-		const keyvRedis = new KeyvRedis();
-		await keyvRedis.set('foo90', 'bar');
-		await keyvRedis.set('foo902', 'bar2');
-		await keyvRedis.set('foo903', 'bar3');
-		await keyvRedis.clear();
-		const value = await keyvRedis.get('foo90');
-		expect(value).toBeUndefined();
-		await keyvRedis.disconnect();
-	});
 
-	test('should clear with no namespace and useUnlink to false', async () => {
-		const keyvRedis = new KeyvRedis();
-		keyvRedis.useUnlink = false;
-		await keyvRedis.set('foo90', 'bar');
-		await keyvRedis.set('foo902', 'bar2');
-		await keyvRedis.set('foo903', 'bar3');
-		await keyvRedis.clear();
-		const value = await keyvRedis.get('foo90');
-		expect(value).toBeUndefined();
-		await keyvRedis.disconnect();
-	});
-
-	test('should clear with no namespace but not the namespace ones', async () => {
-		const keyvRedis = new KeyvRedis();
-		const client = await keyvRedis.getClient() as RedisClientType;
-		await client.flushDb();
-		keyvRedis.namespace = 'ns1';
-		await keyvRedis.set('foo91', 'bar');
-		keyvRedis.namespace = undefined;
-		await keyvRedis.set('foo912', 'bar2');
-		await keyvRedis.set('foo913', 'bar3');
-		await keyvRedis.clear();
-		keyvRedis.namespace = 'ns1';
-		const value = await keyvRedis.get('foo91');
-		expect(value).toBe('bar');
-		await keyvRedis.disconnect();
-	});
-
-	test('should not clear all with no namespace if noNamespaceAffectsAll is false', async () => {
-		const keyvRedis = new KeyvRedis();
-		keyvRedis.noNamespaceAffectsAll = false;
-
-		keyvRedis.namespace = 'ns1';
-		await keyvRedis.set('foo91', 'bar');
-		keyvRedis.namespace = undefined;
-		await keyvRedis.set('foo912', 'bar2');
-		await keyvRedis.set('foo913', 'bar3');
-		await keyvRedis.clear();
-		keyvRedis.namespace = 'ns1';
-		const value = await keyvRedis.get('foo91');
-		expect(value).toBeDefined();
-	});
-
-	test('should clear all with no namespace if noNamespaceAffectsAll is true', async () => {
-		const keyvRedis = new KeyvRedis();
-		keyvRedis.noNamespaceAffectsAll = true;
-
-		keyvRedis.namespace = 'ns1';
-		await keyvRedis.set('foo91', 'bar');
-		keyvRedis.namespace = undefined;
-		await keyvRedis.set('foo912', 'bar2');
-		await keyvRedis.set('foo913', 'bar3');
-		await keyvRedis.clear();
-		keyvRedis.namespace = 'ns1';
-		const value = await keyvRedis.get('foo91');
-		expect(value).toBeUndefined();
-	});
-
-	test('should clear namespace but not other ones', async () => {
-		const keyvRedis = new KeyvRedis();
-		const client = await keyvRedis.getClient() as RedisClientType;
-		await client.flushDb();
-		keyvRedis.namespace = 'ns1';
-		await keyvRedis.set('foo921', 'bar');
-		keyvRedis.namespace = 'ns2';
-		await keyvRedis.set('foo922', 'bar2');
-		await keyvRedis.clear();
-		keyvRedis.namespace = 'ns1';
-		const value = await keyvRedis.get('foo921');
-		expect(value).toBe('bar');
-		await keyvRedis.disconnect();
-	});
-
-	test('should be able to set many keys with namespace', async () => {
-		const keyvRedis = new KeyvRedis('redis://localhost:6379', {namespace: 'ns-many1'});
-		await keyvRedis.setMany([{key: 'foo-many1', value: 'bar'}, {key: 'foo-many2', value: 'bar2'}, {key: 'foo-many3', value: 'bar3', ttl: 5}]);
-		const value = await keyvRedis.get('foo-many1');
-		expect(value).toBe('bar');
-		const value2 = await keyvRedis.get('foo-many2');
-		expect(value2).toBe('bar2');
-		await delay(10);
-		const value3 = await keyvRedis.get('foo-many3');
-		expect(value3).toBeUndefined();
-		await keyvRedis.disconnect();
-	});
-
-	test('should be able to has many keys with namespace', async () => {
-		const keyvRedis = new KeyvRedis('redis://localhost:6379', {namespace: 'ns-many2'});
-		await keyvRedis.setMany([{key: 'foo-has-many1', value: 'bar'}, {key: 'foo-has-many2', value: 'bar2'}, {key: 'foo-has-many3', value: 'bar3', ttl: 5}]);
-		await delay(10);
-		const exists = await keyvRedis.hasMany(['foo-has-many1', 'foo-has-many2', 'foo-has-many3']);
-		expect(exists).toEqual([true, true, false]);
-		await keyvRedis.disconnect();
-	});
-
-	test('should be able to delete many with namespace', async () => {
-		const keyvRedis = new KeyvRedis('redis://localhost:6379', {namespace: 'ns-dm1'});
-		await keyvRedis.setMany([{key: 'foo-delete-many1', value: 'bar'}, {key: 'foo-delete-many2', value: 'bar2'}, {key: 'foo-delete-many3', value: 'bar3', ttl: 5}]);
-		await keyvRedis.deleteMany(['foo-delete-many2', 'foo-delete-many3']);
-		await delay(10);
-		const value = await keyvRedis.get('foo-delete-many1');
-		expect(value).toBe('bar');
-		const value2 = await keyvRedis.get('foo-delete-many2');
-		expect(value2).toBeUndefined();
-		const value3 = await keyvRedis.get('foo-delete-many3');
-		expect(value3).toBeUndefined();
-		await keyvRedis.disconnect();
-	});
-});
 
 describe('KeyvRedis Iterators', () => {
 	beforeEach(async () => {
